Extended logging features
Extended logging features by adding an in-app log viewer (e-mail is not
available in the emulator) and adding information about the environment
newline. Also added a small handler to check whether a new connection is
succesful: earlier on, the 'real' phone returned from a connection on a
failed connection, where the 'emulated' phone didn't return until the
connection was succesful. This caused one issue.